pyproject.toml
[project]
name = "pcb-sim-toolkit"
version = "0.1.0"
description = "Open-source PCB design simulation toolkit using ngspice, KiCad, and OpenEMS"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "numpy>=1.24",
    "matplotlib>=3.7",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1"]

[project.scripts]
pcb-sim = "src.core.cli:main"
pcb-sim-gui = "src.gui.app:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"
